% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{Feature selection via iterative rounds of permuted based feature importance}
\usage{
feature_selection(
  fit_function = NULL,
  data = NULL,
  test = data,
  response = NULL,
  loss_function = NULL,
  stat = stats::median,
  iterations = 1,
  sample_size = NULL,
  sample_frac = NULL,
  predict_function = NULL,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{fit_function}{A function with \code{formula} and \code{data} arguments to fit the
desired model.}

\item{data}{A data to calculate the loss_function.}

\item{test}{A testing data frame to evaluate the loss function. By default
is the data argument.}

\item{response}{Name of the variable response.}

\item{loss_function}{The loss function to evaluate,  Must be a function with 2
arguments: actual and predicted values. Loss function gives a smaller
value if the model have better performance of the model.}

\item{stat}{Default \code{median}. A summary function to compare the values of
the loss of a variable vs full model. If the \code{stat} value of the one variable
is smaller than the value of the loss function full model, then the variable
is removed in that round.}

\item{iterations}{Number of iterations.}

\item{sample_size}{Sample size.}

\item{sample_frac}{Proportion to sample in each iteration.}

\item{predict_function}{Predict function, usually is a function(model, newdata)
which returns a vector (no data frame).}

\item{parallel}{A logical value indicating if the process should be using \code{furrr::future_pmap_dbl}
or \code{purrr::pmap_dbl}.}

\item{...}{Specific arguments for \code{fit_function}.}
}
\description{
Feature selection via iterative rounds of permuted based feature importance
}
