% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance}
\alias{variable_importance}
\title{Variable importance via variable permutations}
\usage{
variable_importance(
  object,
  data = NULL,
  variables = NULL,
  response = NULL,
  metric = NULL,
  iterations = 1,
  sample_size = NULL,
  sample_frac = NULL,
  predict_function = NULL,
  parallel = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{The model.}

\item{data}{A data to calculate the metrics.}

\item{variables}{Variables to use.}

\item{response}{Name of the variable response.}

\item{metric}{The metric to evaluate. Must be a loss function, i.e., greater
value of the function better performance of the model.}

\item{iterations}{Number of iterations.}

\item{sample_size}{Sample size.}

\item{sample_frac}{Proportion to sample in each iteration.}

\item{predict_function}{Predict function, usually is a function(model, newdata)
which returns a vector (no data frame).}

\item{parallel}{A logical value indicating if the process should be using \code{furrr::future_pmap_dbl}
or \code{purrr::pmap_dbl}.}

\item{verbose}{A logical value to indicate if progress bars and other messages
will be showed.}
}
\description{
Variable importance via variable permutations
}
\examples{

lm_model <- lm(mpg ~ ., data = mtcars)

vi <- variable_importance(
  lm_model,
  response = "mpg",
  data = mtcars
)

plot(vi)

}
